Improve the file structure
