+++ lib/scr/thumbnail_viewer.dart
